+++ src/main/kotlin/net/fallingangel/jimmerdto/psi/element/DTODefaultValue.kt
@@ -1,0 +1,16 @@
+package net.fallingangel.jimmerdto.psi.element
+
+import com.intellij.psi.PsiElement
+import net.fallingangel.jimmerdto.psi.mixin.DTOElement
+
+interface DTODefaultValue : DTOElement {
+    val boolean: PsiElement?
+
+    val integer: PsiElement?
+
+    val string: PsiElement?
+
+    val float: PsiElement?
+
+    val `null`: PsiElement?
+}